+++ lib/gasfree_sdk/models/token.rb
@@ -12,19 +12,19 @@ module GasfreeSdk
 
       # @!attribute [r] created_at
       #   @return [Time] When the token was added to GasFree
-      attribute :created_at, Types::JSON::Time
+      attribute :created_at, Types::Time
 
       # @!attribute [r] updated_at
       #   @return [Time] When the token was last updated
-      attribute :updated_at, Types::JSON::Time
+      attribute :updated_at, Types::Time
 
       # @!attribute [r] activate_fee
       #   @return [String] The activation fee in the smallest unit of the token
-      attribute :activate_fee, Types::Amount
+      attribute :activate_fee, Types::String
 
       # @!attribute [r] transfer_fee
       #   @return [String] The transfer fee in the smallest unit of the token
-      attribute :transfer_fee, Types::Amount
+      attribute :transfer_fee, Types::String
 
       # @!attribute [r] supported
       #   @return [Boolean] Whether the token is currently supported